+++ src/SignUpForm.js
@@ -1,412 +1,551 @@
 import React, { Component } from 'react';
 import { Motion, spring } from 'react-motion';
 import user from './user';
-
+import SubmitButton from './components/SubmitButton';
+import {MdArrowForward} from 'react-icons/lib/md';
+import 'md5';
+import './SignUpOrLoginForm.css';
 import './SignUpForm.css';
 
-class ControlButtons extends Component {
-	render() {
-		return (
-			<div className="Control-buttons">
-				<button type="submit" disabled={this.props.disabled} className="SignUpNextButton">
-				{ 'Next' }
-				</button>
-			</div>
-		)
-	}
+class VerifyOtpForm extends Component {
+    state = {
+        status: '',
+        otp: '',
+        done: false,
+        forgotBack: false,
+        disabled: false,
+        checking: false,
+    };
+
+    handleFailed = (err) => {
+        console.log(err);
+
+        this.setState({
+            checking: false,
+            error: true,
+            message: err.message,
+        })
+    };
+
+    handleNext = () => {
+        user.verifyOtp(this.state.otp, this.props.mobile, { onSuccess: (id) => {
+                this.props.done(id)
+                setTimeout(this.props.onSuccess, 1000);
+            },
+            onFailed: this.handleFailed });
+        this.setState({ checking: true });
+    };
+
+    handleChange = ({ target }) => {
+        this.setState({ otp: target.value, error: false });
+    };
+
+
+    render() {
+        return (<div>
+                <h2>Verify OTP</h2>
+                <form className={"SignUpForm"} onSubmit={this.handleNext}>
+                    <div className="SignUpElement">
+                        <p>OTP sent on your registered mobile number.</p>
+                        <div className="StatusMessage">
+                            {this.state.status}
+                        </div>
+                        <div className="Input">
+                            <input
+                                id="otp"
+                                ref = "otp"
+                                disabled={this.state.checking}
+                                type="password"
+                                autoComplete={false}
+                                required
+                                placeholder="Enter OTP"
+                                onChange={this.handleChange}/>
+                        </div>
+                    </div>
+                    <div className={'submitButtonLogin'}>
+                        <button type={'submit'} className={'submitSignUp'} onSubmit={this.handleNext}><MdArrowForward/></button>
+                        {
+                            this.state.error ? <p>{this.state.message}</p> : ""
+                        }
+					</div>
+                </form>
+            </div>
+        );
+    }
 }
 
-class GetName extends Component {
-	state = {
-		name: '',
-	}
+class GetFirstName extends Component {
+    state = {
+        fName: '',
+    }
 
-	handleNext = () => {
- 		this.props.done({ name: this.state.name });
-	}
+    handleNext = () => {
+        this.props.done({ fName : this.state.fName});
+    }
 
-	isValid() {
-		return this.state.name.length > 2;
-	}
+    isValid() {
+        var res = this.state.fName.split(" ");
+        if(res.length > 1){
+            return false;
+        }
+        return this.state.fName.length > 2;
 
-	handleChange = ({ target }) => {
-		this.setState({ name: target.value });
-	}
+    }
+    handleChange = ({target}) => {
+        this.setState({fName : target.value});
+    }
+    get() {
+        return this.state.fName;
+    }
 
-	get() {
-		return this.state.name;
-	}
 
-	render() {
-		return (
-			<div className="SignUpElement">
-				<p className="SignUpElement-description">
-					Your name (will appear on your certificates)
-				</p>
-				<input type="text" name="name" onBlur={this.handleNext} id="username" placeholder={"your name"} onChange={this.handleChange} className="SignUpInput" />
-			</div>
-		)
-	}
+    render() {
+        return (
+
+            <input
+                id="firstName"
+                type="text"
+                autoComplete="false"
+                required
+                name="firstName"
+                placeholder="firstName"
+                onChange={this.handleChange}/>
+
+        )
+    }
+}
+
+class GetLastName extends Component {
+
+    state = {
+        lName : '',
+    }
+    handleNext = () => {
+        this.props.done({ lName : this.state.lName});
+    }
+
+    isValid() {
+        var res = this.state.lName.split(" ");
+        if(res.length > 1){
+            return false;
+        }
+        return this.state.lName.length > 2;
+
+    }
+    handleChange = ({target}) => {
+        this.setState({lName : target.value});
+    }
+    get() {
+        return this.state.lName;
+    }
+
+
+    render() {
+        return (
+            <input
+                id="lastName"
+                type="text"
+                autoComplete="false"
+                required
+                name="lastName"
+                placeholder="lastName"
+                onChange={this.handleChange}/>
+
+        )
+    }
+
 }
 
 class GetNumber extends Component {
-	state = {
-		mobile: '',
-		error: true,
-	}
+    state = {
+        mobile:' ',
+        error: true,
+    }
 
-	get() {
-		return this.state.mobile;
-	}
+    get() {
+        return this.state.mobile;
+    }
 
-	handleNext = ({ target }) => {
-		this.setState({ error: !target.value.match(/[0-9]{10,10}/) })
- 		this.props.done({ mobile: this.state.mobile });
-	}
+    handleNext = ({ target }) => {
+        this.setState({ error: !target.value.match(/[0-9]{10,10}/) })
+        this.props.done({ mobile: this.state.mobile });
+    }
 
-	isValid() {
-		return !this.state.error;
-	}
+    isValid() {
+        return !this.state.error;
+    }
 
-	handleChange = ({ target }) => {
-		this.setState({ mobile: target.value,});
-	}
+    handleChange = ({ target }) => {
+        this.setState({ mobile: target.value,});
+    }
 
-	render() {
-		return (
-			<div className="SignUpElement">
-				<p className="SignUpElement-description">
-				Your mobile number (in 10 digits)
-				</p>
-				<input type="text" name="mobile" onBlur={this.handleNext} placeholder={"9876543210"} className="SignUpInput" onChange={this.handleChange} />
-			</div>
-		)
-	}
+    render() {
+        return (
+            <input
+                id="mobile"
+                type="numeric"
+                autoComplete="false"
+                required
+                name="mobileNumber"
+                placeholder="Contact No."
+                onChange={this.handleChange}/>
+        )
+    }
+
 }
 
 class GetCollege extends Component {
-	state = {
-		college: '',
-	}
+    state = {
+        college: '',
+    };
 
-	handleNext = () => {
- 		this.props.done({ college: this.state.college });
-	}
+    handleNext = () => {
+        this.props.done({ college: this.state.college });
+    };
 
-	get() {
-		return this.state.college;
-	}
+    get() {
+        return this.state.college;
+    }
 
-	isValid() {
-		return this.state.college.length !== 0;
-	}
+    isValid() {
+        return this.state.college.length !== 0;
+    }
 
-	handleChange = ({ target }) => {
-		this.setState({ college: target.value });
-	}
+    handleChange = ({ target }) => {
+        this.setState({ college: target.value });
+    };
 
-	render() {
-		return (
-			<div className="SignUpElement">
-				<p className="SignUpElement-description">
-					Your college (will appear on certificates)
-				</p>
-				<input type="text" name="college" onBlur={this.handleNext} placeholder={"Some awesome college"} className="SignUpInput" onChange={this.handleChange} />
-			</div>
-		)
-	}
+    render() {
+        return (
+            <input
+                id="collegeName"
+                type="text"
+                autoComplete="false"
+                required
+                name = "collegeName"
+                placeholder="College/University"
+                onChange={this.handleChange}/>
+        )
+    }
 }
 
 const emailre = /^(([^<>()\[\]\\.,;:\s@"]+(\.[^<>()\[\]\\.,;:\s@"]+)*)|(".+"))@((\[[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}\.[0-9]{1,3}])|(([a-zA-Z\-0-9]+\.)+[a-zA-Z]{2,}))$/;
 
 class GetEmail extends Component {
-	state = {
-		email: '',
-		error: true,
-	}
+    state = {
+        email: '',
+        error: true,
+    }
 
-	get() {
-		return this.state.email;
-	}
+    get() {
+        return this.state.email;
+    }
 
-	handleNext = ({ target }) => {
-		this.setState({ error: !target.value.match(emailre)})
- 		this.props.done({ email: this.state.email });
-	}
+    handleNext = ({ target }) => {
+        this.setState({ error: !target.value.match(emailre)})
+        this.props.done({ email: this.state.email ,
+            error: !target.value.match(emailre)});
 
-	isValid() {
-		return !this.state.error;
-	}
+    }
 
-	handleChange = ({ target }) => {
-		this.setState({ email: target.value });
-	}
+    handleChange = ({ target }) => {
+        this.setState({ email: target.value });
+    }
 
-	render() {
-		return (
-			<div className="SignUpElement">
-				<p className="SignUpElement-description">
-					Your Email ID
-				</p>
-				<input type="email"
-					name="email"
-					placeholder="example@example.com"
-					className="SignUpInput"
-					onChange={this.handleChange}
-					onBlur={this.handleNext}
-					/>
-			</div>
-		)
-	}
+    render() {
+        return (
+            <input
+                id="emailId"
+                type="email"
+                autoComplete = "off"
+                required
+                name = "emailId"
+                placeholder="example@example.com"
+                onChange={this.handleChange}
+                on/>
+        )
+    }
+}
+
+const pwdre = /^((?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#\$%\^&\*])(?=.{8,}))$/;
+
+class GetPassword extends Component {
+    state = {
+        pwd : '',
+        error: true,
+    }
+
+    get() {
+        return this.state.pwd;
+    }
+
+    handleNext = ( { target }) => {
+        this.setState({ error: !target.value.match(pwdre)});
+        this.props.done({ pwd: this.state.pwd });
+    }
+
+
+    isValid() {
+        return !this.state.error;
+    }
+
+    handleChange = ({ target }) => {
+        this.setState({ pwd: target.value });
+    }
+
+    render() {
+        return (
+            <input
+                id="password"
+                type="password"
+                required
+                name = "password"
+                autoComplete="false"
+                placeholder="Password"
+                onChange={this.handleChange}/>
+        )
+    }
+
 }
 
 class GetGender extends Component {
-	render() {
-		return (
-			<div className="SignUpElement">
-				<p className="SignUpElement-description">
-				What is your gender?
-				</p>
-				<select name="gender" className="SignUpElement-select" onChange={this.props.onChange}>
-					<option value="Male">Male</option>
-					<option value="Female">Female</option>
-					<option value="Other">Other</option>
-				</select>
-			</div>
-		)
-	}
+    render() {
+        return (
+            <div className="InputLabel">
+                <label><input type="radio" name="gender" value="male" required/> Male</label>
+                <label><input type="radio" name="gender" value="female" required/> Female</label>
+                <label><input type="radio" name="gender" value="other" required/> Other</label>
+            </div>
+        )
+    }
 }
 
 class GetAccomodationDetails extends Component {
-	render() {
-		return (
-			<div className="SignUpElement" style={{ marginTop: '2em'}}>
-				<input type="checkbox" name="accomodation" className="SignUpElement-checkbox" onChange={this.props.done} />
-				<label htmlFor="accomodation" className="SignUpElement-description">Do you need accomodation?</label>
-			</div>
-		)
-	}
-}
+    render() {
+        return (
+            <div class="InputLabel">
+                <label><input type="checkbox" name="accomodation" value="yes"/>  Accomodation? </label>
+            </div>
 
-class VerifyMobileNumber extends Component {
-	state = {
-		otp: '',
-		disabled: false,
-		checking: false,
-	}
-
-	handleFailed = (err) => {
-		console.log(err);
-
-		this.setState({
-			checking: false,
-			error: true,
-			message: err.message,
-		})
-	}
-
-	handleNext = () => {
-		user.verifyOtp(this.state.otp, this.props.mobile, { onSuccess: (id) => this.props.done(id), onFailed: this.handleFailed });
-		this.setState({ checking: true });
-	}
-
-	handleChange = ({ target }) => {
-		this.setState({ otp: target.value, error: false });
-	}
-
-	render() {
-		return (
-			<div className="SignUpElement">
-				<div className="SignUpElement-description">
-					<p>OTP has been sent to your mobile.</p>
-				</div>
-				<input type="text"
-					disabled={this.state.checking}
-					placeholder="Enter OTP"
-					className="SignUpInput"
-					onChange={this.handleChange}
-					/>
-
-				<div className="Control-buttons">
-					<button onClick={this.handleNext} disabled={!this.state.otp.length} className="SignUpNextButton">
-					{ 'Next' }
-					</button>
-					{
-						this.state.error ? <p className="SignUpForm-ErrorMessage">{this.state.message}</p> : ""
-					}
-				</div>
-			</div>
-		)
-	}
+        )
+    }
 }
 
 class FinalStep extends Component {
-	render() {
-		return (
-			<div className="FinalStep">
-				<p className="SignUpForm-description">
-					Your PECFEST ID is <span className="pecfestId">{this.props.pecfestId}</span>
-					<br />
-					Now you can login and start registering for events.
-				</p>
-				<div className="Control-buttons">
-					<button className="SignUpNextButton" onClick={() => this.props.done(this.props.pecfestId)}>
-						Continue
-					</button>
-				</div>
-			</div>
-		)
-	}
+    render() {
+        return (
+            <div className="FinalStep">
+                <p className="SignUpForm-description">
+                    Your PECFEST ID is <span className="pecfestId">{this.props.pecfestId}</span>
+                    <br />
+                    Now you can login and start registering for events.
+                </p>
+                <div className="Control-buttons">
+                    <button className="SignUpNextButton" onClick={() => this.props.done(this.props.pecfestId)}>
+                        Continue
+                    </button>
+                </div>
+            </div>
+        )
+    }
 }
 
 
-class SignUpSteps extends Component {
-	state = {
-		done: false,
-		submitting: false,
-		disabled: true,
-		gender: "Male",
-		accomodation: 0,
-	}
-
-	handleDone = (prop) => {
-		const user = Object.assign({}, this.state.user, prop);
-		this.setState({ user, disabled: false });
-	}
-
-	handleSubmit = event => {
-		event.preventDefault();
-
-		const errors = [];
-		if (!this.refs.name.isValid()) {
-			errors.push('Name');
-		}
-
-		if (!this.refs.email.isValid()) {
-			errors.push('Email')
-		}
-
-		if (!this.refs.mobile.isValid()) {
-			errors.push('Mobile')
-		}
-
-		if (!this.refs.college.isValid()) {
-			errors.push('College');
-		}
-
-		if (errors.length > 0) {
-			const message = errors.join(', ') + (errors.length == 1 ? ' is ' : ' are ') + 'invalid.'
-			this.setState({ disabled: true, error: true, message: message });
-			return;
-		}
-
-		const newUser = {
-			name: this.refs.name.get(),
-			email: this.refs.email.get(),
-			mobile: this.refs.mobile.get(),
-			college: this.refs.college.get(),
-			accomodation: this.state.accomodation,
-			gender: this.state.gender
-		}
-
-		this.setState({ user: newUser } );
-
-
-		user.signUp(newUser, {
-			onSuccess: (res) => {
-				this.setState({ submitting: false, otp: true, pecfestId: res.pecfestId });
-			},
-			onFailed: (err) => {
-				if (typeof err.ACK !== 'undefined') {
-					if (err.ACK === 'ALREADY') {
-						this.setState({ message: 'Account already exists. Verifying...' })
-						user.checkVerified(this.state.user.mobile, {
-							onSuccess: verified => {
-								console.log(verified);
-								if (verified) {
-									// send user to the login page
-									this.props.onContinueToLogin()
-								} else {
-									this.setState({ submitting: false, otp: true })
-								}
-							},
-							onFailed: (err) => {
-								this.setState({ submitting: false, error: true, message: 'Unknown error occurred' });
-							}
-						})
-					}
-				}
-				this.setState({ submitting: false, error: true,  disabled: true, message: err.message || 'Some unknown error has occurred.'})
-			}
-		});
-
-		this.setState({ submitting: true, submitMessage: 'Verifying account...' });
-	}
-
-
-	handleAccomo = ({ target }) => {
-		const user = { accomodation: target.checked - 0 };
-		this.setState({ accomodation: user.accomodation, disabled: false })
-	}
-
-
-	handleChange = event => {
-		this.setState({ gender: event.target.value, disabled: false })
-	}
-
-	handleId = id => {
-		this.props.onSignUp(id);
-	}
-
-	renderLoadingOrOtp() {
-		if (this.state.submitting) {
-			return (
-				<div className="SignUpForm-submitting">
-					<p className="SignUpForm-otp-message">{this.state.submitMessage}</p>
-				</div>
-			);
-		}
-
-		// if (this.state.done) {
-			return (
-				<FinalStep pecfestId={this.state.pecfestId} done={this.props.onSignUp} />
-			)
-		// }
-	}
-
-	render() {
-		console.log(this.state);
-		return (
-			<div className="SignUpForm-steps">
-				{
-					this.state.submitting || this.state.otp || this.state.done ?
-						this.renderLoadingOrOtp() :
-						<form autoComplete='off' onSubmit={this.handleSubmit} className="SignUpSteps">
-							{ this.state.error ? <p className="SignUpForm-ErrorMessage">{this.state.message}</p> : ""}
-							<GetName ref="name" done={this.handleDone} />
-							<GetEmail ref="email" done={this.handleDone} />
-							<GetNumber ref="mobile" done={this.handleDone} />
-							<GetCollege ref="college" done={this.handleDone} />
-							<GetGender ref="gender" onChange={this.handleChange} />
-							<GetAccomodationDetails ref="accomodation" done={this.handleAccomo} />
-							<br />
-							<ControlButtons disabled={this.state.disabled} />
-						</form>
-				}
-			</div>
-		)
-	}
-}
 export default class SignUpForm extends Component {
+    constructor(props) {
+        super(props);
+        this.state = {
+            flexState: false,
+            animIsFinished: false,
+            done: false,
+            submitting: false,
+            disabled: true,
+            gender: "Male",
+            accomodation: 0,
+			message:'',
+            otpForm: false,
+
+        };
+        this.fName = React.createRef();
+        this.lName = React.createRef();
+        this.email = React.createRef();
+        this.mobileNumber = React.createRef();
+        this.college = React.createRef();
+        this.gender = React.createRef();
+        this.accomodation = React.createRef();
+        this.password = React.createRef();
+        this.handleSignUp = this.handleSignUp.bind(this);
+    }
+
+
+    isFinished = () => {
+        this.setState({animIsFinished: true});
+    };
+
+    handleDone = (prop) => {
+        const user = Object.assign({}, this.state.user, prop);
+        this.setState({ user, disabled: false });
+    };
+
+    handleLogIn = event => {
+
+    };
+
+    handleVerifyOtp = () => {
+        this.setState({ otp: true })
+    };
+
+    handleSignUp = event => {
+        event.preventDefault();
+
+        const errors = [];
+        if (!this.fName.current.isValid()) {
+            errors.push('firstName');
+        }
+
+        if (!this.lName.current.isValid()) {
+            errors.push('lastName');
+        }
+
+        if (!this.email.current.get().match(emailre)) {
+            errors.push('Email')
+        }
+
+        if (!this.mobileNumber.current.get().match(/[0-9]{10,10}/)) {
+            errors.push('Mobile')
+        }
+
+        if (!this.college.current.isValid()) {
+            errors.push('College');
+        }
+
+        if (errors.length > 0) {
+            const message = errors.join(', ') + (errors.length == 1 ? ' is ' : ' are ') + 'invalid.';
+            this.setState({disabled: true, error: true, message: message});
+            return;
+        }
+        var md5 = require('md5');
+        const newUser = {
+            firstName: this.fName.current.get(),
+            lastName : this.lName.current.get(),
+            email: this.email.current.get(),
+            mobile: this.mobileNumber.current.get(),
+            college: this.college.current.get(),
+            accomodation: this.state.accomodation,
+            gender: this.state.gender,
+            password : md5(this.password.current.get())
+        }
+        this.setState({ user: newUser } );
+
+
+        user.signUp(newUser, {
+            onSuccess: (res) => {
+                user.checkVerified(this.state.user.mobile, {
+                    onSuccess: verified => {
+                        console.log(verified);
+                        if (verified) {
+                            // send user to the login page
+                            //this.props.onContinueToLogin()
+                        } else {
+                            this.setState({ submitting: false, otp: true, pecfestId: res.pecfestId, otpForm: true });
+                        }
+                    },
+                    onFailed: (err) => {
+                        this.setState({ submitting: false, error: true, message: err.message || 'Unknown error occurred.' });
+                    }
+                })
+                this.setState({ submitting: false, error: true,  disabled: true, message: 'Unknown error has occurred.'})
+            },
+            onFailed: (err) => {
+                if (typeof err.ACK !== 'undefined') {
+                    if (err.ACK === 'ALREADY') {
+                        this.setState({ message: 'Account already exists. Verifying...' })
+                        user.checkVerified(this.state.user.mobile, {
+                            onSuccess: verified => {
+                                console.log(verified);
+                                if (verified) {
+                                    // send user to the login page
+                                    //this.props.onContinueToLogin()
+                                } else {
+                                    this.setState({ submitting: false, otp: true, otpForm: true })
+                                }
+                            },
+                            onFailed: (err) => {
+                                this.setState({ submitting: false, error: true, message: err.message || 'Unknown error occurred 2' });
+                            }
+                        })
+                    }
+                }
+                this.setState({ submitting: false, error: true,  disabled: true, message: err.message || 'Some unknown error has occurred. 2'})
+            }
+        });
+
+        this.setState({ submitting: true, submitMessage: 'Verifying account...' });
+    };
+
+
+    handleAccomo = ({ target }) => {
+        const user = { accomodation: target.checked - 0 };
+        this.setState({ accomodation: user.accomodation, disabled: false })
+    };
+
+
+    handleChange = event => {
+        this.setState({ gender: event.target.value, disabled: false })
+    };
+
+    handleId = id => {
+        this.props.onSignUp(id);
+    };
+
+    renderLoadingOrOtp() {
+        if (this.state.submitting) {
+            return (
+                <div className="SignUpForm-submitting">
+                    <p className="SignUpForm-otp-message">{this.state.submitMessage}</p>
+                </div>
+            );
+        }
+
+        // if (this.state.done) {
+        return (
+            <FinalStep pecfestId={this.state.pecfestId} done={this.props.onSignUp} />
+        )
+        // }
+    }
+
 	render() {
+        if (this.state.otpForm) {
+            return (
+				<VerifyOtpForm onSuccess={() => this.setState({ otpForm: false })} mobile={this.mobileNumber.current.get()} />
+
+            )
+        }
+
 		return (
-			<div className="SignUpForm-wrapper">
-				<div className="SignUpForm">
-					<div className="SignUpForm-title">
-						<h1 style={{fontSize: '3.2em'}}>Sign Up</h1>
-					</div>
-					<SignUpSteps onSignUp={this.props.onSignUp} onContinueToLogin={this.props.onContinueToLogin} />
-				</div>
+            <div>
+                <h2>SIGN UP</h2
+                >
+				<p>{this.state.message}</p>
+				<form onSubmit={this.handleSignUp} autoComplete={false}>
+                <div className="Input">
+                    <GetFirstName ref={this.fName} done={this.handleDone}/>
+                    <GetLastName ref={this.lName} done={this.handleDone}/>
+                </div>
+                <div className="Input">
+                    <GetNumber ref={this.mobileNumber} done={this.handleDone}/>
+                    <GetCollege ref={this.college} done={this.handleDone}/>
+                </div>
+                <div className="Input">
+                    <GetEmail ref={this.email} done={this.handleDone}/>
+                    <GetPassword ref={this.password} done={this.handleDone}/>
+                </div>
+                <div className={'Input'}>
+                    <GetGender ref={this.gender} done={this.handleDone}/>
+                    <GetAccomodationDetails ref={this.accomodation} done={this.handleDone}/>
+                </div>
+                <SubmitButton type={this.props.type} disabled={this.props.disabled} onSubmit={this.handleSignUp} onClick={this.handleVerifyOtp}/>
+				</form >
 			</div>
 		)
 	}
